изменения

+++ ewm-main-service/main-server/src/main/java/ru/practicum/model/Event.java
@@ -66,5 +66,5 @@ public class Event {
 
     @Column(name = "views", nullable = false)
     @Builder.Default
-    private Integer views = 0;
+    private Long views = 0L;
 }
